Refactor: Reimplement the application with a modular, service-oriented architecture, introducing new utilities, types, and configurations.
