+++ nodemon.json
@@ -3,6 +3,6 @@
         "dist"
     ],
     "ext": "js",
-    "exec": "node dist/cli.js",
+    "exec": "node dist/bin/github-stats-terminal.js",
     "delay": "200"
 }
